refactor: pending, today, delete, post 쿼리 폴더 -> wakeupSong 쿼리 폴더로 통합
