+++ src/components/Main/TodayMusicList/index.tsx
@@ -1,10 +1,10 @@
 import * as Style from "./style";
 import Title from "components/Common/Title";
+import { useGetTodayMusicDataQuery } from "queries/wakeupSong/wakeupSong.query";
 import { Link } from "react-router-dom";
-import { useGetTodayMusicData } from "queries/todayMusic/todayMusic.query";
 
 const TodayMusicList = () => {
-  const { data } = useGetTodayMusicData();
+  const { data } = useGetTodayMusicDataQuery();
   const nowDate = new Date().getHours();
 
   return (
