--- src/queries/applyMusic/postApplyMusic.query.ts
@@ -1,9 +1,0 @@
-import { useMutation } from "react-query";
-import wakeupSongRepository from "repository/wakeupSong/wakeupSong.repository";
-
-export const usePostApplyMusic = () => {
-  const mutation = useMutation((musicUrl: string) =>
-    wakeupSongRepository.postApplyWakeupSong(musicUrl)
-  );
-  return mutation;
-};
