--- src/queries/deleteMyWakeupSong/deleteMyWakeupSong.query.ts
@@ -1,9 +1,0 @@
-import { useMutation } from "react-query";
-import wakeupSongRepository from "repository/wakeupSong/wakeupSong.repository";
-
-export const useDeleteMyWakeupSongMutation = () => {
-  const mutation = useMutation((musicId: number) =>
-    wakeupSongRepository.deleteMyWakeupSong(musicId)
-  );
-  return mutation;
-};
