--- src/queries/pendingMusic/pendingMusic.query.ts
@@ -1,20 +1,0 @@
-import wakeupSongRepository from "repository/wakeupSong/wakeupSong.repository";
-import { useQuery, UseQueryOptions } from "react-query";
-import { WakeupSongMusicLists } from "types/wakeupSongMy/wakeupSongMy.type.";
-import { AxiosError } from "axios";
-
-export const useGetPendingMusicList = (
-  options?: UseQueryOptions<
-    WakeupSongMusicLists,
-    AxiosError,
-    WakeupSongMusicLists,
-    "pendingMusic/getPendingMusicList"
-  >
-) =>
-  useQuery(
-    "pendingMusic/getPendingMusicList",
-    () => wakeupSongRepository.getWakeupSongPendingMusicListData(),
-    {
-      ...options,
-    }
-  );
